requirements.txt
torch
torchvision

numpy
pandas
tqdm
pytest
pillow

matplotlib
bokeh
pandas-bokeh
seaborn

sacred
pymongo
gitpython

jupyter
jupyterlab
jupytext
